Prevent annotations when logged out.
It appears this was a misuse of `$().ready` that became important when
jQuery was upgraded. In short, `ready` should only be used with
`$(document)` and should not be used to test whether an element on the
page has become ready. In this case, it was being used in hopes of
triggering the callback only when the body contained a class that's
applied for logged in users.