Add basic server side rendering of resources via Vue SSR
Usage via Ruby:

```
resource = Content::Resource.last # grab a Resource with a Case
Vue::SSR.render(resource.resource.content, resource.annotations) # => html string
```

This currently uses webpack to build the ssr.js file on each
invocation, which is inefficient but good for dev. Given that we're
going to be pairing this with PANDOC work, it would make sense to
optimize this with a proper deploy-time build of ssr.js at that point.