Duplicate commonly reference resources
At some point in the past, when a casebook was cloned, this clone was shallow,
and some resources (text blocks and links) were referenced in common between
multiple casebooks. Deleting the resource in one of the casebooks left a
dangling reference in the ContentNodes table. This migration duplicates any
commonly referenced Link or TextBlock.