Add back-end support for tagging casebooks (#1855)
* Add database models to make casebooks taggable

Users would like to be able to tag casebooks with topics from a
controlled vocabulary. This commit creates a Tag model that
serves as the vocabulary and creates a named ManyToMany relationship
between that model and the Casebook model so that casebooks can be
related with many tags.

* Create admin page for casebook tag names

Trusted maintainers should be able to control the tag vocabulary to meet
changing needs. This commit adds a page on the admin site to view all
tag names and add/edit tags.